Update core/components/phpthumbof/model/phpthumbof/phpthumbof.class.php
Updating an image by just "overwriting" it had no effect because the phpThumbOf cache does not care about the date, only about the filename. With this little modification an image can easily be updated by overwriting it without any need to check on all the places this image was used (and refresh the cache for those).